examples/foundary-sdk-python/requirements.txt
azure-ai-agents==1.1.0
azure-ai-projects==1.0.0
azure-core==1.36.0
azure-identity==1.25.1
azure-storage-blob==12.27.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
idna==3.11
isodate==0.7.2
msal==1.34.0
msal-extensions==1.3.1
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.2.1
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.5.0
